Stopping after X revs
The digital crown sets the revolutions before it has started (after that, it sets the speed). Todo is to fix the animation for when isFinished is true (in ContentView).

Also, why do I have to set the text Text("") for it to work?

Plus, is there any less ugly ways then having an if in the update function for elapsedTime?