As far as i can get with drag gesture
I have added stuff from a tutorial (https://www.hackingwithswift.com/books/ios-swiftui/moving-views-with-draggesture-and-offset) but now I cannot figure out how to load the InstructionView() since it is protesting that the result is not used.

Also, sorry for adding it to the same branch as the other fix. Too many merge conflicts in my new branch to solve.